requirements.txt
pypdf
sentence-transformers
faiss-cpu
python-dotenv
openai
streamlit
numpy